requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1
tqdm
PyYAML
numpy
pandas
opencv-python
torch
torchvision
tensorboard
seaborn
matplotlib
scipy
scikit-learn

# backwards compatibility
# pathlib2
